Parsing/kgd-taxpayer/requirements.txt
aioboto3==15.0.0
aiobotocore==2.23.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioitertools==0.12.0
aiosignal==1.4.0
annotated-types==0.7.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
beautifulsoup4==4.13.4
boto3==1.38.27
botocore==1.38.27
fake-useragent==2.2.0
frozenlist==1.7.0
graypy==2.1.0
greenlet==3.2.3
grpcio==1.74.0
grpcio-reflection==1.74.0
grpcio-tools==1.74.0
grpclib==0.4.8
h2==4.2.0
hpack==4.1.0
hyperframe==6.1.0
idna==3.10
jmespath==1.0.1
multidict==6.6.3
propcache==0.3.2
protobuf==6.31.1
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
s3transfer==0.13.1
scrapyx==1.5.20
six==1.17.0
soupsieve==2.7
SQLAlchemy==2.0.41
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
wrapt==1.17.2
yarl==1.20.1
